refactoring/혜택까득 로직 수정
